X,Y:Short => Int

+++ backend/src/main/scala/com/neo/sk/gypsy/gypsyServer/GameServer.scala
@@ -643,7 +643,7 @@ class GameServer(override val boundary: Point) extends Grid {
   override def getGridData(id: String, winWidth: Int, winHeight: Int): GridDataSync = super.getGridData(id, winWidth, winHeight)
 
   def getDataForBot(id:String,winWidth:Int,winHeight:Int):GridData4Bot =  {
-    val currentPlayer = playerMap.get(id).map(a=>(a.x,a.y)).getOrElse(((winWidth/2).toShort,(winHeight/2).toShort ))
+    val currentPlayer = playerMap.get(id).map(a=>(a.x,a.y)).getOrElse(( winWidth/2 ,winHeight/2 ))
     val zoom = playerMap.get(id).map(a=>(a.width,a.height)).getOrElse((30.0,30.0))
     if(getZoomRate(zoom._1,zoom._2,winWidth,winHeight)!=1){
       Scale = getZoomRate(zoom._1,zoom._2,winWidth,winHeight)
